backend/apps/api/urls.py
from django.urls import path
from rest_framework.decorators import api_view
from rest_framework.response import Response

@api_view(['GET'])
def health_check(request):
    """Health check endpoint."""
    return Response({'status': 'healthy'})

urlpatterns = [
    path('health/', health_check, name='health-check'),
] 