backend/requirements.txt
Django==5.0.2
python-dotenv==1.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==8.0.0
pytest-django==4.8.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
dj-database-url==2.1.0 